Change CivicTheme table to CivicTheme component
