+++ docroot/themes/contrib/civictheme/civictheme.libraries.yml
@@ -18,10 +18,3 @@ civictheme.ajax_views:
     - core/drupal
     - core/drupalSettings
     - core/jquery.once
-civictheme.table:
-  version: VERSION
-  js:
-    assets/js/civictheme.table.js: {}
-  dependencies:
-    - core/drupal
-    - core/drupalSettings
